feat: extract tree management into a service.
https://gravitee.atlassian.net/browse/APIM-11813
