+++ gravitee-apim-portal-webui-next/src/app/documentation/services/documentation-tree.service.spec.ts
@@ -1,0 +1,36 @@
+/*
+ * Copyright (C) 2025 The Gravitee team (http://gravitee.io)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import { DocumentationTreeService } from './documentation-tree.service';
+import { MOCK_ITEMS } from '../../../mocks/portal-navigation-item.mocks';
+
+let service: DocumentationTreeService;
+
+beforeEach(() => {
+  service = new DocumentationTreeService();
+});
+
+describe('DocumentationTreeService', () => {
+  it('should create tree', () => {
+    const itemsTree = service.mapItemsToNodes(MOCK_ITEMS);
+
+    expect(itemsTree).toBeTruthy();
+    expect(itemsTree[0].id).toEqual('f1');
+    expect(itemsTree[0].children![0].id).toEqual('f2');
+    expect(itemsTree[0].children![0].children![0].id).toEqual('p1');
+    expect(itemsTree[0].children![1].id).toEqual('p2');
+    expect(itemsTree[1].id).toEqual('p3');
+  });
+});
